ref(compiler): extract TypedCollectionMethodSpecialization from CallSpecialization
Fourth step of splitting the CallSpecialization god-class. Move the typed
persistent-collection accessor lowering (count/nth on PersistentVector,
first/rest on tagged seqs) plus its SEQ_METHODS / SEQ_TAGS tables into a
dedicated TypedCollectionMethodSpecialization collaborator. Now possible
because the shared TagNormalizer landed in the previous step.

Pure move: CallSpecialization dispatch and CallEmitter delegate to the new
class; the existing typed-vector/seq unit tests move alongside it.
Generated PHP is byte-identical (all integration fixtures pass).
