+++ tests/php/Unit/Compiler/Domain/Emitter/OutputEmitter/CallSpecializationTest.php
@@ -15,88 +15,12 @@ use Phel\Compiler\Domain\Emitter\OutputEmitter\CallSpecialization;
 use Phel\Lang\Collections\Map\PersistentMapInterface;
 use Phel\Lang\Collections\Vector\PersistentVectorInterface;
 use Phel\Lang\Keyword;
-use Phel\Lang\SeqInterface;
 use Phel\Lang\Symbol;
 use Phel\Shared\CompilerConstants;
 use PHPUnit\Framework\TestCase;
 
 final class CallSpecializationTest extends TestCase
 {
-    public function test_count_on_typed_vector_specialises_to_method_call(): void
-    {
-        $node = $this->coreCall('count', [$this->vectorLocal('v')]);
-
-        $spec = CallSpecialization::typedVectorMethodCall($node);
-
-        self::assertSame(['method' => 'count', 'args' => []], $spec);
-    }
-
-    public function test_nth_on_typed_vector_specialises_to_get(): void
-    {
-        $node = $this->coreCall('nth', [$this->vectorLocal('v'), new LiteralNode($this->env(), 0)]);
-
-        $spec = CallSpecialization::typedVectorMethodCall($node);
-
-        self::assertSame(['method' => 'get', 'args' => [1]], $spec);
-    }
-
-    public function test_nth_on_untyped_local_falls_back(): void
-    {
-        $env = $this->env();
-        $node = $this->coreCall('nth', [
-            new LocalVarNode($env, Symbol::create('v')),
-            new LiteralNode($env, 0),
-        ]);
-
-        self::assertNull(CallSpecialization::typedVectorMethodCall($node));
-    }
-
-    public function test_count_with_wrong_arity_skips(): void
-    {
-        $node = $this->coreCall('count', [$this->vectorLocal('v'), new LiteralNode($this->env(), 0)]);
-
-        self::assertNull(CallSpecialization::typedVectorMethodCall($node));
-    }
-
-    public function test_other_core_fn_on_typed_vector_skips(): void
-    {
-        $node = $this->coreCall('inc', [$this->vectorLocal('v')]);
-
-        self::assertNull(CallSpecialization::typedVectorMethodCall($node));
-    }
-
-    public function test_first_on_typed_seq_specialises_to_first_method(): void
-    {
-        $node = $this->coreCall('first', [$this->localWithTag('s', SeqInterface::class)]);
-
-        self::assertSame('first', CallSpecialization::typedSeqMethodName($node));
-    }
-
-    public function test_rest_on_typed_vector_specialises_to_rest_method(): void
-    {
-        $node = $this->coreCall('rest', [$this->localWithTag('v', PersistentVectorInterface::class)]);
-
-        self::assertSame('rest', CallSpecialization::typedSeqMethodName($node));
-    }
-
-    public function test_first_on_untyped_local_falls_back(): void
-    {
-        $env = $this->env();
-        $node = $this->coreCall('first', [new LocalVarNode($env, Symbol::create('s'))]);
-
-        self::assertNull(CallSpecialization::typedSeqMethodName($node));
-    }
-
-    public function test_next_is_not_specialised(): void
-    {
-        // `next` returns nil on empty rest; a bare method call cannot
-        // reproduce that without a runtime probe, so the specialiser
-        // stays out of it.
-        $node = $this->coreCall('next', [$this->localWithTag('s', SeqInterface::class)]);
-
-        self::assertNull(CallSpecialization::typedSeqMethodName($node));
-    }
-
     public function test_assoc_on_typed_map_specialises_to_put(): void
     {
         $env = $this->env();
